packages/load-testing/tasks/requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.3
ConfigArgParse==1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-Cors==5.0.0
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.1.1
idna==3.10
importlib-metadata==4.11.3
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.8
MarkupSafe==3.0.0
msgpack==1.1.0
psutil==6.0.0
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
typing_extensions==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.8.0
zope.event==5.0
zope.interface==7.0.3
psycopg2-binary==2.9.9
